Video of Virtual Louvre Museum
Made a virtual Louvre Museum with 8 French artworks and 16 Italian artworks from the Louvre Museum. Worked on the project with Unity version 2021.1.13f1 and Oculus Quest 2. For simple instruction on how to use this project: first, go in front of the artwork you like and point the laser pointer to the artwork. Then, click the artwork which will show you the description of that artwork containing information about the artist, years created, etc... The user can also listen to the audio version of the description when they click the play button. The user can move around the museum by moving the left primary thumbstick of the controller, and by using the right secondary thumbstick the user can rotate the camera viewpoint.